+++ app/src/main/java/com/firebaseapp/gdg_korea_campus/staff/Global.kt
@@ -8,5 +8,4 @@ object Global{
 
     val EVENTLIST_DB_URL = ""
 
-
 }
