+++ app/src/main/java/com/firebaseapp/gdg_korea_campus/staff/view/MainActivity.kt
@@ -1,145 +1,91 @@
 package com.firebaseapp.gdg_korea_campus.staff.view
 
-import android.Manifest
-import android.app.ProgressDialog
-import android.content.Intent
-import android.content.pm.PackageManager
-import android.support.v4.app.ActivityCompat
-import android.support.v4.content.ContextCompat
-import android.support.v7.app.AppCompatActivity
+import android.content.Context
+import android.content.SharedPreferences
 import android.os.Bundle
 import android.support.v7.app.AlertDialog
-import android.util.Base64
+import android.support.v7.app.AppCompatActivity
+import android.support.v7.widget.Toolbar
 import android.util.Log
+import android.view.Menu
+import android.view.MenuItem
+import android.widget.ListView
 import android.widget.Toast
+import com.firebaseapp.gdg_korea_campus.staff.adapter.EventListAdapter
+import com.firebaseapp.gdg_korea_campus.staff.data.source.EventRepository
+import com.firebaseapp.gdg_korea_campus.staff.view.presenter.MainContract
+import com.firebaseapp.gdg_korea_campus.staff.view.presenter.MainPresenter
+
+import android.widget.EditText
 import com.firebaseapp.gdg_korea_campus.staff.R
+import com.firebaseapp.gdg_korea_campus.staff.data.source.PreferenceRepository
 
-import com.google.zxing.integration.android.IntentIntegrator
+/**
+ * Created by lk on 2017. 4. 27..
+ */
+class MainActivity : AppCompatActivity(), MainContract.View {
 
-import org.json.JSONException
-import org.json.JSONObject
+    private lateinit var presenter: MainPresenter
+    private lateinit var toolbar: Toolbar
+    lateinit var eventListAdapter: EventListAdapter
 
-import java.io.IOException
-
-import okhttp3.MultipartBody
-import okhttp3.OkHttpClient
-import okhttp3.Request
-import java.lang.Integer.parseInt
-
-class MainActivity : AppCompatActivity() {
-
-    lateinit var url: String
-    lateinit var mProgressDialog : ProgressDialog
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
+        toolbar = (findViewById(R.id.toolbar) as Toolbar)
+        setSupportActionBar(toolbar)
 
-        url = intent.getStringExtra("url")
+        eventListAdapter = EventListAdapter()
+        (findViewById(R.id.lv_main_eventlist) as ListView).adapter = eventListAdapter
 
-        val permissionCheck = ContextCompat.checkSelfPermission(this, Manifest.permission.CAMERA)
-        if (permissionCheck == PackageManager.PERMISSION_DENIED) {
-            if (ContextCompat.checkSelfPermission(this, Manifest.permission.CAMERA) != PackageManager.PERMISSION_GRANTED) {
+        val preferences = getSharedPreferences("pref", MODE_PRIVATE)
 
-                if (ActivityCompat.shouldShowRequestPermissionRationale(this, Manifest.permission.READ_CONTACTS)) {
-                } else {
-                    ActivityCompat.requestPermissions(this, arrayOf(Manifest.permission.CAMERA), 1)
-                }
-            }
-        } else {
-            IntentIntegrator.initiateScan(this)
+        presenter = MainPresenter().apply {
+            view = this@MainActivity
+            eventData = EventRepository
+            preferenceData = PreferenceRepository(preferences)
+            adapterModel = eventListAdapter
+            adapterView = eventListAdapter
         }
 
+        presenter.loadItems(this, false)
 
     }
 
-    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        val result = IntentIntegrator.parseActivityResult(requestCode, resultCode, data)
-        if (result != null) {
-            if (result.contents == null) {
-                Log.d("MainActivity", "Cancelled scan")
-                Toast.makeText(this, "Cancelled", Toast.LENGTH_LONG).show()
-            } else {
-                Log.d("MainActivity", "${result.contents}")
-
-                val data = decrypt(result.contents, "gdgkrcampus")
-                val d = data.split("/")
-                val thread = NetworkThread(d[0], d[1])
-                thread.start()
-                mProgressDialog = ProgressDialog.show(this@MainActivity,"", "잠시만 기다려 주세요.",true)
-
-            }
-        } else {
-            Log.d("MainActivity", "Weird")
-            // This is important, otherwise the result will not be passed to the fragment
-            super.onActivityResult(requestCode, resultCode, data)
-        }
+    override fun onCreateOptionsMenu(menu: Menu?): Boolean {
+        menuInflater.inflate(R.menu.menu_main, menu)
+        return true
     }
 
-    fun decrypt(data: String, key: String) : String{
-        val result = Base64.decode(data, 0)
-        var results = ""
-        for(i in result.indices){
-            val c = result[i].toInt().let { if(it < 0) it+256 else it }
-            val keychar = key[ (i%key.length -1).let { if (it == -1) key.length-1 else it } ].toInt()
-            results += (c - keychar).toChar()
+    override fun onOptionsItemSelected(item: MenuItem?): Boolean {
+        val id = item?.itemId
+        if(id == R.id.action_settings){
+            presenter.showAPIKeySetDialog(this@MainActivity)
+            return true
         }
-        return results
+        return super.onOptionsItemSelected(item)
+
     }
 
+    override fun showSecuritKeyDialog(_id: String) {
+        Toast.makeText(applicationContext, _id, 1000).show()
+        val ad = AlertDialog.Builder(this@MainActivity)
+        ad.setTitle("비밀키 입력")
 
-    internal inner class NetworkThread(val row: String, val mail: String) : Thread() {
+        val et = EditText(this@MainActivity)
+        ad.setView(et)
 
-        override fun run() {
-            val client = OkHttpClient()
-            val jsonObject = JSONObject()
-            try {
-                jsonObject.put("row", parseInt(row))
-                jsonObject.put("mail", mail)
-            } catch (e: JSONException) {
-                e.printStackTrace()
-            }
-
-            val body1 = MultipartBody.Builder().setType(MultipartBody.FORM)
-                    .addFormDataPart("row", "${parseInt(row) + 1}")
-                    .addFormDataPart("mail", mail)
-                    .build()
-            val request = Request.Builder()
-                    .url(url)
-                    .post(body1)
-                    .build()
-            try {
-                val response = client.newCall(request).execute()
-                val responseString = response.body().string().toString()
-                Log.e("Result", responseString)
-                runOnUiThread {
-                    mProgressDialog.dismiss()
-                    if (responseString == "{\"result\":\"success\"}") {
-                        val alert = AlertDialog.Builder(this@MainActivity)
-                        alert.setPositiveButton("OK"){ dialog, which ->
-                             dialog.dismiss()
-                        }
-                        alert.setMessage("확인 되었습니다")
-                        alert.setOnDismissListener {
-                            IntentIntegrator.initiateScan(this@MainActivity)
-                        }
-                        alert.show()
-                    } else {
-                        val alert = AlertDialog.Builder(this@MainActivity)
-                        alert.setPositiveButton("OK"){ dialog, which ->
-                            dialog.dismiss()
-                        }
-                        alert.setMessage("확인되지 않은 사용자입니다.")
-                        alert.show()
-                    }
-                }
-
-            } catch (e: IOException) {
-                e.printStackTrace()
-            }
-
+        ad.setPositiveButton("확인") { dialog, which ->
+            val value = et.text.toString()
+            Log.e("value", value + "")
+            presenter.loadOpenKey(this@MainActivity,_id,value)
+            dialog.dismiss()
         }
+
+        ad.setNegativeButton("취소") { dialog, which ->
+            dialog.dismiss()
+        }
+        ad.show()
     }
-
-
 }
